Show annotations based on decoder id
Originally, decoder id was equal to decoder instance id, so it didn't
matter what you were comparing with.  However, with distinct decoder
instances, this assumption fails.  Because instance ids can't be known
in advance, and because it would make the CLI arguments even more
complex, we simply let the -A option continue to be per decoder, not per
instance.  This maintains compatibility with prior behaviour, but
doesn't allow the full flexibility possible with the library.

Signed-off-by: Karl Palsson <karlp@etactica.com>